+++ utils/rfcx-checkin/checkin-audio.js
@@ -10,7 +10,6 @@ const audioUtils = require('../../utils/rfcx-audio').audioUtils
 const assetUtils = require('../../utils/internal-rfcx/asset-utils.js').assetUtils
 
 const cachedFiles = require('../../utils/internal-rfcx/cached-files.js').cachedFiles
-const aiService = require('../../services/legacy/ai/ai-service')
 
 const moment = require('moment-timezone')
 const urls = require('../../utils/misc/urls')
@@ -220,40 +219,6 @@ exports.audio = {
             }
           })
     })
-  },
-
-  queueForTaggingByActiveV3Models: function (audioInfo, dbGuardian) {
-    return aiService.getPublicAis({ isActive: true })
-      .then((ais) => {
-        return ais.filter((ai) => {
-          return ai.guardiansWhitelist && ai.guardiansWhitelist.length && ai.guardiansWhitelist.includes(dbGuardian.guid)
-        })
-      })
-      .then((ais) => {
-        const promises = []
-        ais.forEach((ai) => {
-          const name = aiService.combineTopicQueueNameForGuid(ai.guid)
-          const message = {
-            guid: audioInfo.audio_guid,
-            guardian_guid: dbGuardian.guid,
-            guardian_shortname: dbGuardian.shortname,
-            site_guid: dbGuardian.Site.guid,
-            site_timezone: dbGuardian.Site.timezone,
-            measured_at: audioInfo.dbAudioObj.measured_at,
-            file_extension: audioInfo.dbAudioObj.Format.file_extension,
-            capture_sample_count: audioInfo.dbAudioObj.capture_sample_count,
-            sample_rate: audioInfo.dbAudioObj.Format.sample_rate,
-            latitude: dbGuardian.latitude,
-            longitude: dbGuardian.longitude
-          }
-          const prom = aws.publish(name, message)
-            .then((data) => {
-              return data
-            })
-          promises.push(prom)
-        })
-        return Promise.all(promises)
-      })
   },
 
   rollBackCheckIn: function (audioInfo) {